v2-fuzz: do away with the submodule approach for fuzzers
