--- fuzz/README.md
@@ -1,3 +1,0 @@
-# Fuzz Submodule
-
-At the time of writing, [go-fuzz isn't compatible with Go modules](https://github.com/dvyukov/go-fuzz/issues/195). This workaround is adapted from [mvdan's](https://github.com/mvdan) work as described in [this comment](https://github.com/dvyukov/go-fuzz/issues/195#issuecomment-523526736) and implemented [here](https://github.com/mvdan/sh/commit/6c13161a56af5dece4e33497e909ee4cbe3ee6bf).
